refactor code to call `getgetRenderableTimeEstimate` method once
