+++ ui/app/components/app/transaction-list/transaction-list.container.js
@@ -6,13 +6,33 @@ import {
   nonceSortedCompletedTransactionsSelector,
   nonceSortedPendingTransactionsSelector,
 } from '../../../selectors/transactions'
-import { getSelectedAddress, getAssetImages } from '../../../selectors/selectors'
+import { getSelectedAddress, getAssetImages, getSelectedToken } from '../../../selectors/selectors'
 import { selectedTokenSelector } from '../../../selectors/tokens'
 import { updateNetworkNonce } from '../../../store/actions'
 import { fetchBasicGasAndTimeEstimates, fetchGasEstimates } from '../../../ducks/gas/gas.duck'
-import { getBasicGasEstimateBlockTime } from '../../../selectors/custom-gas'
+import {
+  getBasicGasEstimateBlockTime,
+  getCustomGasPrice,
+  getEstimatedGasPrices,
+  getEstimatedGasTimes,
+  formatTimeEstimate,
+  getFastPriceEstimateInHexWEI,
+} from '../../../selectors/custom-gas'
+import { getAdjacentGasPrices, extrapolateY } from '../gas-customization/gas-price-chart/gas-price-chart.utils'
+import { hexWEIToDecGWEI } from '../../../helpers/utils/conversions.util'
 
-const mapStateToProps = state => {
+const mapStateToProps = (state, ownProps) => {
+  const { selectedAddressTxList } = state.metamask
+  const { modalState: { props: modalProps } = {} } = state.appState.modal || {}
+  const { transaction = {} } = ownProps
+  const { txData = {} } = modalProps || {}
+  const selectedTransaction = selectedAddressTxList.find(({ id }) => id === (transaction.id || txData.id))
+  const { gasPrice: currentGasPrice } = getTxParams(state, selectedTransaction)
+  const customModalGasPriceInHex = getCustomGasPrice(state) || currentGasPrice
+  const customGasPrice = calcCustomGasPrice(customModalGasPriceInHex)
+  const gasPrices = getEstimatedGasPrices(state)
+  const estimatedTimes = getEstimatedGasTimes(state)
+
   return {
     completedTransactions: nonceSortedCompletedTransactionsSelector(state),
     pendingTransactions: nonceSortedPendingTransactionsSelector(state),
@@ -20,6 +40,8 @@ const mapStateToProps = state => {
     selectedAddress: getSelectedAddress(state),
     assetImages: getAssetImages(state),
     blockTime: getBasicGasEstimateBlockTime(state),
+    customGasPrice,
+    currentTimeEstimate: getRenderableTimeEstimate(customGasPrice, gasPrices, estimatedTimes),
   }
 }
 
@@ -47,3 +69,47 @@ export default compose(
   withRouter,
   connect(mapStateToProps, mapDispatchToProps, mergeProps)
 )(TransactionList)
+
+function getRenderableTimeEstimate (currentGasPrice, gasPrices, estimatedTimes) {
+    const minGasPrice = gasPrices[0]
+    const maxGasPrice = gasPrices[gasPrices.length - 1]
+    let priceForEstimation = currentGasPrice
+    if (currentGasPrice < minGasPrice) {
+        priceForEstimation = minGasPrice
+    } else if (currentGasPrice > maxGasPrice) {
+        priceForEstimation = maxGasPrice
+    }
+
+    const {
+        closestLowerValueIndex,
+        closestHigherValueIndex,
+        closestHigherValue,
+        closestLowerValue,
+    } = getAdjacentGasPrices({ gasPrices, priceToPosition: priceForEstimation })
+
+    const newTimeEstimate = extrapolateY({
+        higherY: estimatedTimes[closestHigherValueIndex],
+        lowerY: estimatedTimes[closestLowerValueIndex],
+        higherX: closestHigherValue,
+        lowerX: closestLowerValue,
+        xForExtrapolation: priceForEstimation,
+    })
+
+    return formatTimeEstimate(newTimeEstimate, currentGasPrice > maxGasPrice, currentGasPrice < minGasPrice)
+}
+
+function calcCustomGasPrice (customGasPriceInHex) {
+    return Number(hexWEIToDecGWEI(customGasPriceInHex))
+}
+
+function getTxParams (state, selectedTransaction = {}) {
+    const { metamask: { send } } = state
+    const { txParams } = selectedTransaction
+    return txParams || {
+        from: send.from,
+        gas: send.gasLimit || '0x5208',
+        gasPrice: send.gasPrice || getFastPriceEstimateInHexWEI(state, true),
+        to: send.to,
+        value: getSelectedToken(state) ? '0x0' : send.amount,
+    }
+}
